Remove all reflection
This commit has two changes:

1. Update the project.clj to use `:global-vars`
2. Fix a type hint to remove reflection

For the first change, the previous usage of `:warn-on-reflection` is not
deprecated. For the second change, we had an incorrect type hint of
`Node` when using the `.select` method. `Node` doesn't have that method,
but `Element` does. Reflection was covering this for us, but that breaks
when using reaver within a Clojure project compiled to a Graal native
image.

You can see both of these issues, on master, by using this:

```
❯ lein check
WARNING: :warn-on-reflection is deprecated in project.clj; use :global-vars.
Compiling namespace reaver
Reflection warning, reaver.clj:102:26 - call to method select on org.jsoup.nodes.Node can't be resolved (no such method).
```

Now, with these changes:

```
❯ lein check
Compiling namespace reaver
```